requirements.txt
great-expectations
pandas
pyarrow
python-telegram-bot==20.7
